labels on proposal and proposals list
