+++ apps/ui/src/types.ts
@@ -257,6 +257,7 @@ export type Proposal = {
   max_end: number;
   snapshot: number;
   choices: string[];
+  labels: string[];
   scores: number[];
   scores_total: number;
   execution_time: number;
