+++ apps/ui/src/views/Proposal.vue
@@ -320,6 +320,9 @@ watchEffect(() => {
               :decimals="votingPowerDecimals"
             />
           </div>
+          <div v-if="space.labels?.length && proposal.labels?.length">
+            <ProposalLabels :proposal="proposal" :labels="space.labels" />
+          </div>
           <div>
             <h4 class="mb-2.5 eyebrow flex items-center gap-2">
               <IH-clock />
